Experimental Commit - Referencing Unity's HappyHarvest Example
This is a somewhat unstable commit of importing some of Unity's HappyHarvest's scripts/prefabs to study and reference.

This is from Unity's own official example project on making a 2D top down game. The features looked at are a Sunrise and Moonrise cycle, and the dependencies it has with a gamemanager.