Added Bone Dog glow, and much more
Outside of the visuals, I introduced Unity Jobs into our project. I refactored some of the code/calculations to now run multi-threaded.

I also addressed the Tooltip bug which would get caught on UI bits from lifting the pointer caused by the drag and drop script.

I began working on the "Attack" system used for all creatures and refactored some of the actions done by the AI. This includes some bits of awareness and the old Action "Chase" to be "Hunt" now. Chase is just a method for now which will be a helper method in the Creature AI as other behaviors may want to "Chase" something.